requirements.txt
# Xử lý tiếng Việt
pyvi>=0.1.1

# Xử lý embedding văn bản, hình ảnh, audio
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
pillow>=9.0.0
soundfile>=0.12.0

# Xử lý ảnh
opencv-python>=4.8.0

# Xử lý audio
librosa>=0.10.0
pydub>=0.25.0

# FAISS cho vector search
faiss-cpu>=1.7.0

# API backend
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# HTTP client cho testing
requests>=2.31.0

# Data processing
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Utilities
tqdm>=4.65.0